Added starterkit theme
