--- .npmrc
@@ -1,3 +1,0 @@
-@fortawesome:registry=https://npm.fontawesome.com/
-@awesome.me:registry=https://npm.fontawesome.com/
-//npm.fontawesome.com/:_authToken=${FONTAWESOME_PACKAGE_TOKEN}
