+++ css/baseVars.css
@@ -1,0 +1,15 @@
+/* Base variables */
+
+:root {
+  --theme-primary: #fff;
+  --theme-secondary: #F4F4F5;
+  --theme-brand: #047857;
+  --theme-brand-hover: #064E3B;
+  --theme-border: #E4E4E7;
+  --theme-success: #F0FDF4;
+  --theme-success-border: #22C55E;
+  --theme-warning: #FEFCE8;
+  --theme-warning-border: #FCD34D;
+  --theme-error: #FEF2F2;
+  --theme-error-border: #DC2626;
+}
