+++ js/baseIcons.js
@@ -1,0 +1,23 @@
+// Create arrays with the icon names and add additional icons
+export const baseFaBrandIcons = [
+  "faXTwitter"
+];
+
+export const baseFaSolidIcons = [
+  "faBars",
+  "faXmark",
+  "faSun",
+  "faMoon",
+  "faMagnifyingGlass",
+  "faAngleDown",
+  "faUser",
+  "faSignOut",
+];
+
+export const baseFaRegularIcons = [
+  "faWindowClose",
+];
+
+export const baseFaProRegularIcons = [
+  "faCircleInfo"
+];
