--- prettier.config.mjs
@@ -1,7 +1,0 @@
-// prettier.config.mjs
-
-/** @type {import("prettier").Config} */
-const config = {
-};
-
-export default config;
