+++ themes/starterkit_project_theme/.gitignore
@@ -1,0 +1,2 @@
+node_modules
+.npmrc
