+++ themes/starterkit_project_theme/assets/css/components/base-overrides.css
@@ -1,6 +1,4 @@
-@import "baseVars.css";
-
-@layer base-components {
+@layer base-overrides {
   /* Define common background colors */
 
   .theme-bg-primary {
@@ -141,9 +139,4 @@
   .alert-danger {
     @apply bg-red-100 text-red-700  border-red-300;
   }
-}
-
-/* Set relative container to avoid site hiding behind admin sidebar */
-.dialog-off-canvas-main-canvas {
-  position: relative;
 }
