+++ themes/starterkit_project_theme/assets/css/components/content.css
@@ -1,3 +1,7 @@
+/**
+ * HTML content usually generated by editors in wysiwyg.
+ */
+
 .content {
 
   h2 {
