+++ themes/starterkit_project_theme/assets/css/components/tailwind.css
@@ -1,0 +1,3 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
