+++ themes/starterkit_project_theme/assets/css/styles.css
@@ -1,0 +1,4 @@
+@import "../../../../contrib/itkdev_base_theme/css/styles.css";
+@import "components/content.css";
+@import "components/print.css";
+@import "components/project-styles.css";
