+++ themes/starterkit_project_theme/assets/js/components/project-icons.js
@@ -1,3 +1,4 @@
+// Import icon names from the base theme
 export const projectFaBrandIcons = [];
 export const projectFaSolidIcons = [];
 export const projectFaRegularIcons = [];
