+++ themes/starterkit_project_theme/assets/js/components/project-scripts.js
