--- themes/starterkit_project_theme/css/project-styles.css
@@ -1,74 +1,0 @@
-@import "print.css";
-@import "content.css";
-
-/* Project specific styles */
-:root {
-    --theme-primary: #fff;
-    --theme-secondary: #F4F4F5;
-    --theme-brand: #047857;
-    --theme-brand-hover: #064E3B;
-    --theme-border: #E4E4E7;
-    --theme-success: #F0FDF4;
-    --theme-success-border: #22C55E;
-    --theme-warning: #FEFCE8;
-    --theme-warning-border: #FCD34D;
-    --theme-error: #FEF2F2;
-    --theme-error-border: #DC2626;
-}
-
-.alert-success {
-  border-color: var(--theme-success-border);
-  background-color: var(--theme-success);
-}
-
-.alert-warning {
-  border-color: var(--theme-warning-border);
-  background-color: var(--theme-warning);
-}
-
-.alert-danger {
-  border-color: var(--theme-error-border);
-  background-color: var(--theme-error);
-}
-
-.tablesort {
-  &.desc {
-    height: 100%;
-    vertical-align: text-bottom;
-  }
-
-  &.asc {
-    height: 100%;
-    vertical-align: sub;
-  }
-}
-
-/**
- * @file
- *
- * Replaces standard throbber from stable9 with Olivero's.
- *
- * See https://www.drupal.org/docs/develop/theming-drupal/update-the-throbber
- *
- * Copied and edited from:
- * web/core/themes/olivero/css/components/autocomplete-loading.module.css
- *
- */
-
-/* remove ugly spinner from autocomplete input  */
-.js input.form-autocomplete {
-  background-image: none;
-}
-
-/* replace spinner, and doing this it only shows when on input and load */
-.js input.form-autocomplete.ui-autocomplete-loading {
-  background-image: url("data:image/svg+xml,%3csvg xmlns='http://www.w3.org/2000/svg' width='18' height='18' viewBox='0 0 10 10'%3e  %3cstyle type='text/css'%3e%40keyframes s%7b0%25%7btransform:rotate(0deg) translate(-50%25,-50%25)%7d50%25%7btransform:rotate(430deg) translate(-50%25,-50%25);stroke-dashoffset:20%7d100%25%7btransform:rotate(720deg) translate(-50%25,-50%25)%7d%7dellipse%7banimation:s 1s linear infinite%7d%3c/style%3e  %3cg transform='translate(5 5)'%3e    %3cellipse fill='none' ry='4' rx='4' cy='5' cx='5' stroke='%237e96a7' stroke-width='1' stroke-dashoffset='6.125' stroke-dasharray='25' transform='translate(-5 -5)'/%3e  %3c/g%3e%3c/svg%3e");
-  background-position: 98% center;
-}
-
-/* Style dropdown from autocomplete input */
-.ui-autocomplete .ui-menu-item {
-  height: 50px;
-  border-bottom: 1px solid rgb(203 213 225); /* bg-slate-300 */
-  color: rgb(71 85 105);
-}
