+++ themes/starterkit_project_theme/starterkit_project_theme.info.yml
@@ -11,3 +11,5 @@ regions:
   navigation: Navigation
   content: Content
   footer: Footer
+libraries:
+  - starterkit_project_theme/styles
