+++ themes/starterkit_project_theme/starterkit_project_theme.starterkit.yml
@@ -6,5 +6,5 @@ info:
   version: 1.0.0
 starterkit:
   replace_patterns:
-    - 'itkdev_project_theme'
+    - 'starterkit_project_theme'
